pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "openai-agents[litellm]>=0.6.2",
    "openai-chatkit<=1.4.0",
    "python-dotenv==1.0.1",
    "uvicorn[standard]==0.32.1",
]
